Port FileInputStreamTest
